test: isolate insertion regressions from sibling fixes
